1) Removed more windows warnings. 2) Cleanup of old unused, and non-bullet code 3) Fixed wheelies, though they are currently still    disabled in the stk_config.data file since the    AI crashes with the new wheelie implementation.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@1365 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/gui/font.cpp
@@ -79,7 +79,7 @@ void Font::Print(const char *text, int size,
     int   sz          = (int)(size*std::max(scale_x,scale_y)*fontScaling);
 
     float l,r,t,b;
-    m_fnt->getBBox(text, sz, 0, &l, &r, &b, &t);
+    m_fnt->getBBox(text, (float)sz, 0, &l, &r, &b, &t);
     const int W = (int)((r-l+0.99));
     const int H = (int)((t-b+0.99));
 
@@ -104,7 +104,7 @@ void Font::Print(const char *text, int size,
     }
 
     m_text_out->begin();
-    m_text_out->setPointSize(sz);
+    m_text_out->setPointSize((float)sz);
     if(doShadow)
     {
         m_text_out->start2f((GLfloat)x-2, (GLfloat)y-2);
